Separate determination and creation of require infra
