+++ server/commands/deployments/DeployService.js
@@ -108,7 +108,8 @@ function validateCommandAndCreateDeployment(command) {
 function deploy(deployment, destination, sourcePackage, command) {
   return co(function* () {
     let accountName = deployment.accountName;
-    yield provideInfrastructure(accountName, deployment, command);
+    let requiredInfra = yield getInfrastructureRequirements(accountName, deployment, command);
+    yield provideInfrastructure(accountName, requiredInfra, command);
     yield preparePackage(accountName, destination, sourcePackage, command);
     yield pushDeployment(accountName, deployment, destination, command);
 
@@ -140,11 +141,22 @@ function sanitiseError(error) {
   return _.toString(error);
 }
 
-function provideInfrastructure(accountName, deployment, parentCommand) {
+function getInfrastructureRequirements(accountName, deployment, parentCommand) {
+  let command = {
+    name: 'GetInfrastructureRequirements',
+    accountName,
+    deployment
+  };
+
+  return sender.sendCommand({ command, parent: parentCommand });
+}
+
+function provideInfrastructure(accountName, requiredInfra, parentCommand) {
   let command = {
     name: 'ProvideInfrastructure',
     accountName,
-    deployment
+    asgsToCreate: requiredInfra.asgsToCreate,
+    launchConfigsToCreate: requiredInfra.launchConfigsToCreate
   };
 
   return sender.sendCommand({ command, parent: parentCommand });
